Add files via upload
this is a initial UML design, not final for DD1. Changes are going to be made to it. 
Possibly remove the specific property classes (e.g. ENV, SCI, MKV, UWP, etc) and use an index based system to identify each space on the board. Also possibly remove the Goose, GRTbus, THDonut, etc classes and just use strings. 
playerOptions() is for doing something on the property i.e. buy, mortgage , rent, etc. performEvent() in Game calls playerOptions() for the current player. 
The structure of the classes is not final either, I wasn't sure how to structure it but atm it seems ok? We should probably implement an observer class but  i think we can get away without making one tbh. 
Controller is basically meant to start the game. It makes a new GameBoard, etc. It's basically the main.cc file/user interface.
BoardSpace is supposed to the current space the player is on. 
I'm still not sure how we should structure the property/non-property classes.